pyproject.toml
# Team Invotex — Muzamil Shaik
# Meta x PyTorch x HuggingFace OpenEnv Hackathon 2026

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-invotex"
version = "0.1.0"
description = "OpenEnv multi-domain RL environment: Email Triage, Traffic Control, Customer Support"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "sqlalchemy>=2.0.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "prometheus-client>=0.20.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.0",
]
train = [
    "accelerate>=0.34.0",
    "datasets>=2.20.0",
    "peft>=0.12.0",
    "transformers>=4.44.0",
    "trl>=0.10.1",
]

[tool.setuptools]
include-package-data = true
packages = [
    "openenv_invotex",
    "openenv_invotex.server",
    "openenv_invotex.server.utils",
    "openenv_invotex.domains",
    "openenv_invotex.domains.email_triage",
    "openenv_invotex.domains.traffic_control",
    "openenv_invotex.domains.customer_support",
]
package-dir = { "openenv_invotex" = "." }
